Added in a table for the players that sorts them by their positions. -There may be a bug with this and react and the rendering that happens if you refresh on the teams page.
Tried to work on some stuff related to the players. Could not get it to fetch from the NHL API when trying to grab the ID in the url to call the fetch request before loading the page. Need to look further into this.

NOTE: ONLY CLICKING ON MIKKO RANTANEN FROM THE COLORADO AVALANCHE WORKS AT THIS VERY MOMENT

Added in players headshot photos.